requirements.txt
# eace - Easy ECE Dependencies
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
google-generativeai>=0.8.3
pydantic>=2.10.0
